changed single inference.py to two separate entry point and pipeline
